feat: migration for cascading + delete on adm page

+++ node-api/src/features/user/user.service.ts
@@ -29,6 +29,12 @@ export class UserService {
   }
 
   async delete(id: number): Promise<User> {
+    const userExists = await this.getById(id);
+
+    if (!userExists) {
+      throw new NotFound('El usuario no existe');
+    }
+
     try {
       return await this.prisma.user.delete({ where: { id } });
     } catch (error) {
